+++ frontend/src/models/levels/tools/_selection.ts
@@ -452,6 +452,11 @@ export const $openedSelectionEdit = createStore<string | null>(null)
 // REFACT: flush old content (if any) to level? then put new content to level?
 $drawState.on(submitSelectionEdit, (s, r) => {
   if (s?.op === Op.STABLE) {
-    return { ...s, content: r };
+    return {
+      ...s,
+      width: r.tiles.width,
+      height: r.tiles.height,
+      content: r,
+    };
   }
 });