+++ frontend/src/models/ui/hotkeys-defined.ts
@@ -11,6 +11,7 @@ export const HK_EDIT_MAZE: HotKeyShortcuts = [["M", HotKeyMask.SHIFT]];
 export const HK_EDIT_RANDOM: HotKeyShortcuts = [["N", HotKeyMask.SHIFT]];
 export const HK_EDIT_REPLACE: HotKeyShortcuts = [["R", HotKeyMask.SHIFT]];
 export const HK_EDIT_SHUFFLE: HotKeyShortcuts = [["S", HotKeyMask.SHIFT]];
+export const HK_EDIT_TRANSPOSE: HotKeyShortcuts = [["T", HotKeyMask.SHIFT]];
 export const HK_PASTE: HotKeyShortcuts = [["V", HotKeyMask.CTRL], ["Paste"]];
 export const HK_REDO: HotKeyShortcuts = [["Y", HotKeyMask.CTRL], ["Redo"]];
 export const HK_TOOL_FLOOD: HotKeyShortcuts = [["F"]];